reorganized console-specific handling. added startup version check
