+++ Binner/Library/Binner.Common/Security/CertificateStatus.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Binner.Common.Security
+{
+    /// <summary>
+    /// Status of certificate registration
+    /// </summary>
+    [Flags]
+    public enum CertificateState
+    {
+        None,
+        Created,
+        Registered,
+        Error
+    }
+}
